Fix devise edit password action. Also refactor some routing, user management and related specs.
