+++ Gemfile
@@ -4,8 +4,8 @@ gem 'ancestry', '~> 2.1.0'
 gem 'cancancan', '~> 1.13.1'
 gem 'cocoon', '~> 1.2.6'
 gem 'coffee-rails', '~> 4.1.0'
-gem 'devise', '~> 3.5', '>= 3.5.10'
-gem 'devise-i18n', '~> 0.12.1'
+gem 'devise'
+gem 'devise-i18n'
 gem 'faker', '~> 1.5.0'
 gem 'foundation-rails', '= 5.5.1.0'
 gem 'friendly_id', '~> 5.1.0'
