+++ app/views/admin/users/_manage_fields.html.erb
