+++ app/views/admin/users/show.html.erb
