+++ app/views/devise/passwords/edit.html.erb
@@ -1,0 +1,45 @@
+<div class="row">
+  <div class="small-12 columns">
+    <h2 class="title-l"><%= t('main.title')%></h2>
+  </div>
+</div>
+
+<div class="row">
+  <div class="small-12 medium-6 medium-centered columns">
+    <h2><%= t('.change_your_password') %></h2>
+
+    <%= form_for(resource, as: resource_name, url: password_path(resource_name), html: { method: :put }) do |f| %>
+      <%= devise_error_messages! %>
+
+      <% if resource.errors.any? %>
+        <div class="panel callout radius">
+          <strong><%= t('activerecord.errors.template.header', model: user_by_role(resource), count: resource.errors.count) %></strong>
+          <ul>
+            <% resource.errors.full_messages.each do |msg| %>
+                <li><%= msg %></li>
+            <% end %>
+          </ul>
+        </div>
+      <% end %>
+
+      <%= f.hidden_field :reset_password_token %>
+
+      <div class="form-group">
+        <%= f.label :password %>
+        <%= f.password_field :password, autocomplete: "off" %>
+      </div>
+
+      <div class="form-group">
+        <%= f.label :password_confirmation %>
+        <%= f.password_field :password_confirmation, autocomplete: "off" %>
+      </div>
+
+      <div class="actions">
+        <%= f.submit t('.change_my_password'), class: 'button right' %>
+      </div>
+    <% end %>
+
+    <%= render "devise/shared/links" %>
+
+  </div>
+</div>
