+++ app/views/devise/passwords/new.html.erb
@@ -7,14 +7,14 @@
 <div class="row">
   <div class="small-12 medium-6 medium-centered columns">
     <%= devise_error_messages! %>
-    <%= form_for(resource, :as => resource_name, :url => password_path(resource_name), :html => { :method => :post, :role => 'form'}) do |f| %>
-      <h3><%= t'devise.forgot_password' %></h3>
-      <p><%= t'devise.password_reset_instructions' %></p>
+    <%= form_for(resource, as: resource_name, url: password_path(resource_name), html:  { method: :post, role: 'form'} ) do |f| %>
+      <h3><%= t '.forgot_your_password' %></h3>
+      <p><%= t '.send_me_reset_password_instructions' %></p>
       <div class="form-group">
-        <%= f.label t('devise.email') %>
-        <%= f.email_field :email, :autofocus => true, class: 'form-control' %>
+        <%= f.label :email %>
+        <%= f.email_field :email, autofocus: true, class: 'form-control' %>
       </div>
-       <%= f.submit t('devise.reset_password'), :class => 'button right' %>
+       <%= f.submit t('.send_me_reset_password_instructions'), class: 'button right' %>
     <% end %>
   </div>
 </div>
