--- app/views/devise/registrations/edit.html.erb
@@ -1,22 +1,0 @@
-<h1><%= t("backend.change_password") %></h1>
-
-<%= form_for(resource, as: resource_name, url: user_password_registration_path(), html: { method: :put, class: 'admin-form'}) do |f| %>
-  <%= devise_error_messages! %>
-
-  <fieldset>
-    <legend><%= t('backend.modify') %></legend>
-
-    <div class="field">
-      <%= f.label(:password, t("backend.new_password")) %>
-      <%= f.password_field :password, autocomplete: "off" %>
-    </div>
-
-    <div class="field">
-      <%= f.label(:password_confirmation, t("backend.new_password_confirmation")) %>
-      <%= f.password_field :password_confirmation, autocomplete: "off" %>
-    </div>
-
-  </fieldset>
-
-  <%= f.submit "#{t("backend.save")}", class: "button radius success right" %>
-<% end %>
