+++ app/views/devise/sessions/new.html.erb
@@ -6,35 +6,32 @@
 
 <div class="row">
   <div class="small-12 medium-6 medium-centered columns">
-    <%= form_for(resource, :as => resource_name, :url => session_path(resource_name), :html => { :role => 'form'}) do |f| %>
+    <%= form_for(resource, as: resource_name, url: session_path(resource_name), html: { role: 'form' }) do |f| %>
       <div>
         ¿No tienes una cuenta? <%= link_to 'Regístrate', "https://sede.madrid.es", target: "_blank" %>
       </div>
       <br />
       <%= devise_error_messages! %>
       <div class="form-group">
-        <%- if devise_mapping.registerable? %>
-          <%= link_to 'Sign up', new_registration_path(resource_name), class: 'right'  %>
-        <% end -%>
-        <%= f.label t('devise.email') %>
-        <%= f.email_field :email, :autofocus => true, class: 'form-control' %>
+        <%= f.label :email %>
+        <%= f.email_field :email, autofocus: true, class: 'form-control' %>
       </div>
       <div class="form-group">
-        <%= f.label t('devise.password') %>
-        <%= f.password_field :password, class: 'form-control' %>
+        <%= f.label :password %>
+        <%= f.password_field :password, autocomplete: "off", class: 'form-control' %>
         <%- if devise_mapping.recoverable? %>
-          <%= link_to t('devise.forgot_password') , new_password_path(resource_name), class: 'right' %>
-        <% end -%>
+          <%= link_to t(".forgot_your_password"), new_password_path(resource_name), class: 'right' %>
+        <% end %>
       </div>
 
       <% if devise_mapping.rememberable? -%>
-        <div class="checkbox" style="width:250px">
-          <label>
-            <%= f.check_box :remember_me %> <%= t'devise.remember' %>
-          </label>
+        <div class="field">
+          <%= f.check_box :remember_me %>
+          <%= f.label :remember_me %>
         </div>
       <% end -%>
-      <%= f.submit t('devise.sign_in'), :class => 'button' %>
+
+      <%= f.submit t('.sign_in'), class: 'button' %>
     <% end %>
   </div>
 </div>
