+++ config/application.rb
@@ -20,6 +20,8 @@ module Agendas
       g.fixture_replacement :factory_girl, dir: "spec/factories"
     end
 
+    config.to_prepare { Devise::Mailer.layout "mailer" }
+
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
