+++ config/locales/devise.es.yml
@@ -1,72 +1,8 @@
-# Traducciones adicionales en https://github.com/plataformatec/devise/wiki/I18n
-
 es:
   devise:
-    sign_in: "Entrar"
-    email: "Correo electrónico"
-    password: "Contraseña"
-    forgot_password: "¿Olvidaste tu contraseña?"
-    remember: "Recordar en este ordenador"
-    password_reset_instructions: "Te enviaremos instrucciones de restablecimiento de contraseña"
-    reset_password: "Restablecer la contraseña"
-    confirmations:
-      confirmed: "Tu cuenta ha sido confirmada."
-      send_instructions: "Recibirás un correo electrónico en unos minutos con instrucciones sobre cómo restablecer tu contraseña."
-      send_paranoid_instructions: "Si tu correo electrónico existe en nuestra base de datos recibirás un correo electrónico en unos minutos con instrucciones sobre cómo restablecer tu contraseña."
-    failure:
-      already_authenticated: "Ya has iniciado sesión."
-      inactive: "Tu cuenta aún no ha sido activada."
-      invalid: "%{authentication_keys} o contraseña inválidos."
-      locked: "Tu cuenta ha sido bloqueada."
-      last_attempt: "Tienes un último intento antes de que tu cuenta sea bloqueada."
-      not_found_in_database: "%{authentication_keys} o contraseña inválidos."
-      timeout: "Tu sesión ha expirado, por favor inicia sesión nuevamente para continuar."
-      unauthenticated: "Necesitas iniciar sesión o registrarte para continuar."
-      unconfirmed: "Debes confirmar tu cuenta para continuar."
-    mailer:
-      confirmation_instructions:
-        subject: "Instrucciones de confirmación"
-      unlock_instructions:
-        subject: "Instrucciones de desbloqueo"
-      reset_password_instructions:
-        action: Cambiar mi contraseña
-        greeting: "¡Hola %{recipient}!"
-        instruction: Alguien ha solicitado un enlace para cambiar su contraseña, lo que se puede realizar a través del siguiente enlace.
-        instruction_2: Si usted no lo ha solicitado, por favor ignore este correo electrónico.
-        instruction_3: Su contraseña no será cambiada hasta que usted acceda al enlace y cree una nueva.
-        subject: Instrucciones de recuperación de contraseña
-    omniauth_callbacks:
-      failure: "No se te ha podido autorizar de %{kind} debido a \"%{reason}\"."
-      success: "Identificado correctamente de %{kind}."
-    passwords:
-      no_token: "No puedes acceder a esta página si no es a través de un enlace para restablecer la contraseña. Si has accedido desde el enlace para restablecer la contraseña, asegúrate de que la URL esté completa."
-      send_instructions: "Recibirás un correo electrónico con instrucciones sobre cómo restablecer tu contraseña en unos minutos."
-      send_paranoid_instructions: "Si tu correo electrónico existe en nuestra base de datos, recibirás un enlace para restablecer la contraseña en unos minutos."
-      updated: "Tu contraseña ha cambiado correctamente. Has sido identificado correctamente."
-      updated_not_active: "Tu contraseña se ha cambiado correctamente."
-    registrations:
-      destroyed: "¡Adiós! Tu cuenta ha sido cancelada. Esperamos volver a verte pronto."
-      signed_up: "¡Bienvenido! Has sido identificado."
-      signed_up_but_inactive: "Te has registrado correctamente, pero no has podido iniciar sesión porque tu cuenta no ha sido activada."
-      signed_up_but_locked: "Te has registrado correctamente, pero no has podido iniciar sesión porque tu cuenta está bloqueada."
-      signed_up_but_unconfirmed: "Se te ha enviado un mensaje con un enlace de confirmación. Por favor visita el enlace para activar tu cuenta."
-      update_needs_confirmation: "Has actualizado tu cuenta correctamente, sin embargo necesitamos verificar tu nueva cuenta de correo. Por favor revisa tu correo electrónico y visita el enlace para finalizar la confirmación de tu nueva dirección de correo electrónico."
-      updated: "Has actualizado tu cuenta correctamente."
     sessions:
-      signed_in: "Has iniciado sesión correctamente."
-      signed_out: "Has cerrado la sesión correctamente."
-      already_signed_out: "Has cerrado la sesión correctamente."
-    unlocks:
-      send_instructions: "Recibirás un correo electrónico en unos minutos con instrucciones sobre cómo desbloquear tu cuenta."
-      send_paranoid_instructions: "Si tu cuenta existe, recibirás un correo electrónico en unos minutos con instrucciones sobre cómo desbloquear tu cuenta."
-      unlocked: "Tu cuenta ha sido desbloqueada. Por favor inicia sesión para continuar."
-  errors:
-    messages:
-      already_confirmed: "ya has sido confirmado, por favor intenta iniciar sesión."
-      confirmation_period_expired: "necesitas ser confirmado en %{period}, por favor vuelve a solicitarla."
-      expired: "ha expirado, por favor vuelve a solicitarla."
-      not_found: "no se ha encontrado."
-      not_locked: "no estaba bloqueado."
-      not_saved:
-        one: "1 error impidió que este %{resource} fuera guardado:"
-        other: "%{count} errores impidieron que este %{resource} fuera guardado:"
+      new:
+        forgot_your_password: ¿Olvidaste la contraseña?
+    failure:
+      not_found_in_database: "%{authentication_keys} o contraseña inválidos."
+      invalid: "%{authentication_keys} o contraseña inválidos."
