+++ config/routes.rb
@@ -27,10 +27,7 @@ Rails.application.routes.draw do
   get "/admin", to: 'events#index', as: 'admin'
 
   devise_for :users, controllers: { sessions: "users/sessions" }
-  get 'admin/edit_password', to: 'admin/passwords#edit', as: 'edit_password'
-  match 'admin/update_password', to: 'admin/passwords#update', as: 'update_password', via: [:patch, :put]
 
-  resources :users
   resources :events
   resources :areas
   resources :activities
@@ -38,6 +35,7 @@ Rails.application.routes.draw do
   resources :holders
 
   namespace :admin do
+    resources :users
     resources :organizations do
       resources :agents, except: :show
       resources :organization_interests, only: [:index]
