--- spec/features/users/sign_in_spec.rb
@@ -1,42 +1,0 @@
-feature 'Sign in', :devise do
-
-  scenario 'user cannot sign in if not registered' do
-    signin('test@example.com', 'please123')
-    expect(page).to have_content I18n.t 'devise.failure.not_found_in_database', authentication_keys: 'email'
-  end
-
-  scenario 'user can sign in with valid credentials' do
-    user = FactoryGirl.create(:user)
-    signin(user.email, user.password)
-    expect(page).to have_content I18n.t 'devise.sessions.signed_in'
-  end
-
-  scenario 'user cannot sign in with wrong email' do
-    user = FactoryGirl.create(:user)
-    signin('invalid@email.com', user.password)
-    expect(page).to have_content I18n.t 'devise.failure.not_found_in_database', authentication_keys: 'email'
-  end
-
-  scenario 'user cannot sign in with wrong password' do
-    user = FactoryGirl.create(:user)
-    signin(user.email, 'invalidpass')
-    expect(page).to have_content I18n.t 'devise.failure.invalid', authentication_keys: 'email'
-  end
-
-  scenario 'render link to register sede.electonica' do
-    visit admin_path
-
-    expect(page).to have_content "¿No tienes una cuenta?"
-    expect(page).to have_content "¿Olvidaste tu contraseña?"
-    expect(page).to have_link("Regístrate", href: "https://sede.madrid.es")
-  end
-
-  scenario 'should not allow soft deleted user' do
-    user = FactoryGirl.create(:user, deleted_at: DateTime.current)
-
-    signin(user.email, user.password)
-
-    expect(page).to have_content "Tu cuenta ha sido bloqueada."
-  end
-
-end
