+++ spec/support/helpers/session_helpers.rb
@@ -12,7 +12,7 @@ module Features
       visit new_user_session_path
       fill_in 'user_email', with: email
       fill_in 'user_password', with: password
-      click_button 'Entrar'
+      click_button 'Iniciar sesión'
     end
   end
 end
